Fixed problem with gradlew execution permission denied, fixed problem with 2 package names, minor refactoring
